--- src/components/CheckPopUp/CheckPopUp.tsx
@@ -1,1 +1,0 @@
-export const checkId = (item: number) => item > 3;
