+++ src/components/GameOverPopUp/GameOverPopUp.tsx
@@ -1,4 +1,3 @@
-// import { Link } from 'react-router-dom';
 import './losepopup.scss';
 
 type GameOverProps = {
