+++ src/components/GameOverPopUp/losepopup.scss
