+++ src/pages/playersPage/PlayerSettings.tsx
@@ -4,9 +4,9 @@ import { Link } from 'react-router-dom';
 import { PlayerType } from './PlayerSettings-interface';
 import { Counter } from '../../components/Counter/Counter';
 import { Player } from '../../components/Players/Player';
-import './player-settings.scss';
 import { Context } from '../../App';
 import { heroes } from '../../data/heroes';
+import './player-settings.scss';
 
 const maxPlayers = 4;
 const title = 'Выберите количество игроков:';
@@ -14,8 +14,7 @@ const list = 'Список игроков';
 const startText = 'Начать игру';
 
 export const PlayerSettings = () => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { play, changePlayers } = useContext(Context);
+  const { changePlayers } = useContext(Context);
   const startplayers: PlayerType[] = [
     { id: 1, name: 'Player1', isHuman: true, hero: heroes[0] },
   ];
